Add check for canceled request in Axios useEffect
When the useEffect() is run in dev environments, it actually runs twice. This can cause the abort controller to interrupt the operation and gets caught in the catch on line 22. Adding an if statement prevents the abort from punting the user out to the login page. I also added axiosPrivate, isLoading, navigate, and location to the dependency array since they are defined outside the useEffect and it was causing the following warning:

React Hook useEffect has missing dependencies: 'axiosPrivate', 'location', 'navigate' and 'location'. Either include them or remove the dependency array  react-hooks/exhaustive-deps.

Now with these changes, I can set isLoading to true and force a re-render and fetch of the data, such as when closing out a modal or other user driven action that might want to cause a refresh of the data.